code review updates following TL meeting
