+++ Enigmatry.Entry.AspNetCore.Authorization/Attributes/UserHasRoleAttribute.cs
@@ -6,12 +6,15 @@ using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Logging;
 
 namespace Enigmatry.Entry.AspNetCore.Authorization.Attributes;
+
 public sealed class UserHasRoleAttribute : AuthorizeAttribute, IAuthorizationFilter
 {
     private const string PolicyPrefix = "UserHasRole";
 
+    // Ienumerable
     public new string[] Roles
     {
+        // Move to constructor
         get => Policy != null ? Policy[PolicyPrefix.Length..].Split(',') : Array.Empty<string>();
         private init => Policy = $"{PolicyPrefix}{String.Join(',', value)}";
     }
@@ -20,11 +23,12 @@ public sealed class UserHasRoleAttribute : AuthorizeAttribute, IAuthorizationFil
     /// Initializes a new instance of the <see cref="UserHasRoleAttribute"/> class.
     /// </summary>
     /// <param name="roles">A comma delimited list of roles that are allowed to access the resource.</param>
-    public UserHasRoleAttribute(string roles) : base(PolicyNames.UserHasRole)
+    public UserHasRoleAttribute(string roles) : base(PolicyPrefix)
     {
-        Roles = roles.Split(',');
+        Roles = roles.Split(',', StringSplitOptions.TrimEntries);
     }
 
+    // Constructor with params strings
 
     public void OnAuthorization(AuthorizationFilterContext context)
     {
