+++ Enigmatry.Entry.AspNetCore.Authorization/AuthorizationStartupExtensions.cs
@@ -1,9 +1,12 @@
 ﻿using Enigmatry.Entry.AspNetCore.Authorization.Requirements;
+using JetBrains.Annotations;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Enigmatry.Entry.AspNetCore.Authorization;
+
+[PublicAPI]
 public static class AuthorizationStartupExtensions
 {
     public static ControllerActionEndpointConventionBuilder AppRequireAuthorization(
