--- Enigmatry.Entry.AspNetCore.Authorization/PolicyNames.cs
@@ -1,7 +1,0 @@
-﻿namespace Enigmatry.Entry.AspNetCore.Authorization;
-public static class PolicyNames
-{
-    public const string UserAuthenticated = "UserAuthenticated";
-    public const string UserHasPermission = "UserHasPermission";
-    public const string UserHasRole = "UserHasRole";
-}
