+++ Enigmatry.Entry.AspNetCore.Authorization/Requirements/UserHasRoleRequirement.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Authorization;
+
+namespace Enigmatry.Entry.AspNetCore.Authorization.Requirements;
+
+internal class UserHasRoleRequirement : IAuthorizationRequirement
+{
+    public UserHasRoleRequirement(string roles)
+    {
+        Roles = roles;
+    }
+
+    // Array!
+    public string Roles { get; }
+}
